+++ src/duties/index.js
@@ -1,3 +1,4 @@
+let _ = require('lodash/fp')
 let pr = require('./pr')
 let lint = require('./lint')
 let testResults = require('./testResults')
@@ -5,7 +6,7 @@ let library = require('./library')
 let autoFix = require('./autoFix')
 let browserResults = require('./browserResults')
 
-module.exports = Object.assign(
+let api = Object.assign(
   {},
   pr,
   lint,
@@ -14,3 +15,4 @@ module.exports = Object.assign(
   library,
   autoFix
 )
+module.exports = globals => _.mapValues(f => f(globals), api)