+++ src/duties/lint.js
@@ -13,32 +13,32 @@ let formatLintMessages = _.flow(
   _.join('\n')
 )
 
-let lintTemplate = (severity, danger, config) => lint => `
+let lintTemplate = (severity, danger) => (lint, rootFolder) => `
 <details>
   <summary>${basename(lint.filePath)}</summary>
   <p>${linkifyPath({
     danger,
-    path: pathTail(lint.filePath, config),
+    path: pathTail(lint.filePath, rootFolder),
   })}</p>
   ${formatLintMessages(
     severity ? _.filter({ severity }, lint.messages) : lint.messages
   )}
 </details>`
 
-let formatLint = (severity, danger, config) =>
+let formatLint = (severity, danger) => (lintResults, rootFolder) =>
   _.flow(
     _.filter(lint => _.some({ severity }, lint.messages)),
-    _.map(lintTemplate(severity, danger, config)),
+    _.map(v => lintTemplate(severity, danger)(v, rootFolder)),
     _.join('')
-  )
+  )(lintResults)
 
-let formatStandardJSLint = (danger, config) =>
+let formatStandardJSLint = danger => (lintResults, rootFolder) =>
   _.flow(
-    _.map(lintTemplate(undefined, danger, config)),
+    _.map(v => lintTemplate(undefined, danger)(v, rootFolder)),
     _.join('')
-  )
+  )(lintResults)
 
-let hasLintErrors = ({ lintResults, fail, message, danger, config }) => {
+let hasLintErrors = ({ fail, message, danger }) => (lintResults, rootFolder = '.') => {
   if (_.isNil(lintResults)) {
     message('Could not find any lint results')
     return
@@ -52,15 +52,15 @@ let hasLintErrors = ({ lintResults, fail, message, danger, config }) => {
   if (isStandardOutput) {
     if (lintResults && lintResults.length > 0) {
       fail(`Your PR has lint errors. Please fix these and commit them.
-        ${formatStandardJSLint(danger, config)(lintResults)}`)
+        ${formatStandardJSLint(danger)(lintResults, rootFolder)}`)
     }
   } else if (_.sumBy('errorCount', lintResults) > 0) {
     fail(`Your PR has lint errors. Please fix these and commit them.
-      ${formatLint(2, danger, config)(lintResults)}`)
+      ${formatLint(2, danger)(lintResults, rootFolder)}`)
   }
 }
 
-let hasLintWarnings = ({ lintResults, warn, message, danger, config }) => {
+let hasLintWarnings = ({ warn, message, danger }) => (lintResults, rootFolder = '.') => {
   if (_.isNil(lintResults)) {
     message('Could not find any lint results')
     return
@@ -68,7 +68,7 @@ let hasLintWarnings = ({ lintResults, warn, message, danger, config }) => {
   if (_.sumBy('warningCount', lintResults) > 0) {
     warn(
       `Your PR has lint warnings. Please consider fixing these.
-      ${formatLint(1, danger, config)(lintResults)}`
+      ${formatLint(1, danger)(lintResults, rootFolder)}`
     )
   }
 }